adicionando rotas e atualizando o swagger
